Allow gophersay to be called as a library
also some collaborative refactors with @adamryman with our go learns over the years
